Finished changes
I think I don't want to continue working on this anymore, there are a few things to improve, but this is already too much.

I continued to work on this because I want to evolve (fork) this project beyond Pong and make a simple engine, this is why app.c is way too complex for a game like Pong.

I think this is all I can code for Pong and it's time to move on.